+++ trading_system/patterns/__init__.py
@@ -1,12 +1,22 @@
 """Chart-pattern scan + multi-timeframe bias."""
 
-from trading_system.patterns.detectors import DetectedPattern, best_reversal, scan_patterns
+from trading_system.patterns.detectors import (
+    DetectedPattern,
+    WEAK_PATTERNS,
+    best_reversal,
+    measure_target,
+    pattern_opposes_htf,
+    scan_patterns,
+)
 from trading_system.patterns.mtf import combine_htf_votes, ltf_turn, macd_htf_bias, resample_ohlcv
 from trading_system.patterns.pivots import find_pivots, is_near
 
 __all__ = [
     "DetectedPattern",
+    "WEAK_PATTERNS",
     "best_reversal",
+    "measure_target",
+    "pattern_opposes_htf",
     "scan_patterns",
     "combine_htf_votes",
     "ltf_turn",